event_service/requirements.txt
fastapi==0.63.0
bson==0.5.10
click==7.1.2
dataclasses==0.8
decorator==4.4.2
dnspython==2.1.0
h11==0.12.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.22.1
pydantic==1.7.3
pymongo==3.11.2
python-dateutil==2.8.1
python-decouple==3.4
shortuuid==1.0.1
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.3
validators==0.18.2
gunicorn
httptools 
uvloop
redis-om==0.0.20
